+++ src/app/page.tsx
@@ -15,14 +15,14 @@ export default function Home() {
       <div className='flex flex-col gap-4 w-full '>
         <Link
           href={ROUTE.SIGN_UP}
-          className='py-4 bg-blue-500 w-full rounded-lg text-white text-xl font-bold text-center'
+          className='py-4 bg-blue-500 w-full rounded-lg text-white text-l font-bold text-center'
         >
           <i className='ri-mail-line mr-1' />
           <span>이메일로 회원가입</span>
         </Link>
         <Link
           href={ROUTE.SIGN_IN}
-          className='py-4 w-full rounded-lg text-black border border-zinc-500 text-xl font-bold text-center'
+          className='py-4 w-full rounded-lg text-black border border-zinc-500 text-l font-bold text-center'
         >
           로그인
         </Link>
